kr3/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
slowapi
python-dotenv
pydantic-settings
